+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 22 September 2023
+
+### Added
+
+* Added `short_name` and `address` fields to the company information for use in report templates (Closes #339)
 
 ## [4.0.0] - 20 September 2023